feat(ingest): implement message storage and reconciliation pipeline
