+++ packages/lib/src/email/email-storage.ts
@@ -1,6 +1,12 @@
 // packages/lib/src/email/email-storage.ts
+//
+// Legacy compatibility shim. All ingest logic lives in `packages/lib/src/ingest/`;
+// this file keeps the `MessageStorageService` class alive for existing callers
+// that instantiate it directly (providers, webhook routes, tests).
+//
+// New code should use `createIngestContext` + the functional exports from
+// `@auxx/lib/ingest` instead of this class.
 
-import { database as db, schema } from '@auxx/database'
 import {
   EmailLabel,
   IdentifierType as IdentifierTypeEnum,
@@ -8,1716 +14,141 @@ import {
   ParticipantRole as ParticipantRoleEnum,
   ThreadStatus,
 } from '@auxx/database/enums'
-import type {
-  IdentifierType,
-  MessageEntity as Message,
-  ParticipantEntity as Participant,
-  ParticipantRole,
-  ThreadEntity as Thread,
-} from '@auxx/database/types'
-import { createScopedLogger } from '@auxx/logger'
-import { and, eq, inArray, isNull, sql } from 'drizzle-orm'
-import { v4 as uuidv4 } from 'uuid'
-import { SelectiveModeCache } from '../cache/selective-mode-cache'
-import { linkContactToCompanyByDomain } from '../ingest/companies/link-contact'
-import { getOwnDomains } from '../ingest/domain/classifier'
-import { MessageReconcilerService } from '../messages/message-reconciler.service'
-import { ThreadManagerService } from '../messages/thread-manager.service'
-import { UnifiedCrudHandler } from '../resources/crud/unified-handler'
-import { SystemUserService } from '../users/system-user-service'
+import type { MessageEntity as Message, ThreadEntity as Thread } from '@auxx/database/types'
+import {
+  batchStoreMessages,
+  createContactAfterOutboundMessage,
+  createIngestContext,
+  deleteMessagesByExternalIds,
+  ensureContactsForRecipients,
+  getThread,
+  getThreadMessages,
+  type IngestContext,
+  type IntegrationSettings,
+  storeMessage,
+} from '../ingest'
 
-const logger = createScopedLogger('message-storage')
+// Type re-exports (MessageData etc. live in ingest/types now)
+export type {
+  IntegrationSettings,
+  MessageAttachmentMeta,
+  MessageData,
+  ParticipantInputData,
+} from '../ingest/types'
 
-/**
- * Check if an email matches a filter entry (exact email or domain match).
- */
-function matchesFilterEntry(email: string, entries: string[]): boolean {
-  const normalized = email.toLowerCase().trim()
-  const domain = normalized.split('@')[1]
-  return entries.some((entry) => normalized === entry || domain === entry)
-}
-
-// Re-export enums for convenience
+// Enum re-exports preserved for backwards compat.
 export {
-  MessageType,
   EmailLabel,
   IdentifierTypeEnum as IdentifierType,
+  MessageType,
   ParticipantRoleEnum as ParticipantRole,
   ThreadStatus,
 }
-// JSON types
-type JsonValue = any
-type JsonArray = any
-
-// --- Interfaces ---
-
-// Structure for participant info provided by conversion methods
-export interface ParticipantInputData {
-  identifier: string // The raw identifier (email, phone, PSID, etc.)
-  name?: string | null // Match Drizzle type
-  raw?: string | null // Original raw string if available
-}
 
 /**
- * Provider-side attachment metadata for downstream ingest (not persisted in DB).
- */
-export interface MessageAttachmentMeta {
-  filename: string
-  mimeType: string
-  size: number
-  inline: boolean
-  contentId: string | null
-  /** Gmail: body.attachmentId for large parts; null for embedded parts */
-  providerAttachmentId?: string | null
-  /** Gmail: base64url-encoded body.data for small embedded parts */
-  embeddedData?: string | null
-}
-
-// Structure for message data coming from provider conversion methods
-export interface MessageData {
-  externalId: string
-  externalThreadId: string
-  inboxId?: string
-  integrationId: string
-  organizationId: string // Essential context
-
-  // Core message fields
-  isInbound: boolean
-  subject?: string | null
-  textHtml?: string | null
-  textPlain?: string | null
-  snippet?: string | null
-  metadata?: JsonValue | null
-  createdTime: Date
-  sentAt: Date
-  receivedAt: Date
-
-  // Participant Data
-  from: ParticipantInputData // Required
-  to: ParticipantInputData[] // Required, can be empty
-  cc?: ParticipantInputData[]
-  bcc?: ParticipantInputData[]
-  replyTo?: ParticipantInputData[]
-
-  // Attachments
-  hasAttachments: boolean
-
-  // Object-backed body storage (set by ingest pipeline)
-  htmlBodyStorageLocationId?: string | null
-
-  // Optional/Provider-specific fields
-  historyId?: number | null // DB schema uses bigint({ mode: 'number' })
-  internetMessageId?: string | null
-  keywords?: string[]
-  labelIds?: string[] // Raw provider label IDs
-  inReplyTo?: string | null
-  references?: string | null
-  threadIndex?: string | null
-  folderId?: string | null
-  internetHeaders?: JsonArray | null
-  isAutoReply?: boolean | null
-  isFirstInThread?: boolean | null // Provider might determine this
-  isAIGenerated?: boolean | null
-
-  /** Provider-side attachment metadata for downstream ingest (not persisted in DB) */
-  providerAttachments?: MessageAttachmentMeta[]
-}
-
-/**
- * Interface for integration settings
- */
-interface IntegrationSettings {
-  recordCreation?: {
-    mode: 'all' | 'selective' | 'none'
-  }
-  excludeSenders?: string[]
-  excludeRecipients?: string[]
-  onlyProcessRecipients?: string[]
-  [key: string]: any
-}
-
-/**
- * Service class responsible for storing Messages, Threads, Participants,
- * and related data into the database based on the revised schema.
+ * Thin compatibility wrapper around the functional ingest pipeline.
+ *
+ * Behaviour:
+ * - Each call lazily builds (and caches) an `IngestContext` — so you can
+ *   still do `new MessageStorageService(orgId)` and call `storeMessage`
+ *   repeatedly without re-running system-user resolution.
+ * - Callers that construct without an organizationId pull it from
+ *   `messageData.organizationId` on first call (webhook handlers).
+ * - `setIntegrationSettings` / `setInitialSyncMode` write onto the cached
+ *   context, preserving existing call-site ergonomics.
  */
 export class MessageStorageService {
+  private readonly ctxByOrg = new Map<string, Promise<IngestContext>>()
   private integrationSettings?: IntegrationSettings
-  private selectiveCache: SelectiveModeCache
   private isInitialSync = false
-  private reconciler?: MessageReconcilerService
-  private threadManager?: ThreadManagerService
-  private organizationId?: string
-  /** Cached UnifiedCrudHandler instances per organization for batch performance */
-  private crudHandlerCache = new Map<string, UnifiedCrudHandler>()
-  /** Cached system user IDs per organization */
-  private systemUserIdCache = new Map<string, string>()
-  /** Per-batch cache of domain → companyId for auto-company linking (dedupes within a batch) */
-  private companyIdByDomain = new Map<string, string | null>()
-  /** Per-batch cache of org → own-domains set (avoids redundant org-cache reads) */
-  private ownDomainsByOrg = new Map<string, Set<string>>()
+  private readonly defaultOrganizationId?: string
 
   constructor(organizationId?: string) {
-    this.selectiveCache = new SelectiveModeCache()
-    if (organizationId) {
-      this.organizationId = organizationId
-      this.reconciler = new MessageReconcilerService(
-        organizationId,
-        new ThreadManagerService(organizationId, db),
-        db
-      )
-      this.threadManager = new ThreadManagerService(organizationId, db)
-    }
+    this.defaultOrganizationId = organizationId
   }
 
-  /**
-   * Check if a message should be ignored based on integration filter settings.
-   *
-   * Evaluation order:
-   * 1. onlyProcessRecipients (allowlist) — if set and no TO matches → ignore
-   * 2. excludeSenders (blocklist) — if FROM matches → ignore
-   * 3. excludeRecipients (blocklist) — skipped if onlyProcessRecipients already passed
-   */
-  private shouldIgnoreMessage(messageData: MessageData): boolean {
-    const settings = this.integrationSettings
-    if (!settings) return false
-
-    const fromEmail = messageData.from?.identifier?.toLowerCase().trim()
-    const toEmails = (messageData.to ?? [])
-      .map((p) => p.identifier?.toLowerCase().trim())
-      .filter(Boolean) as string[]
-    const ccEmails = (messageData.cc ?? [])
-      .map((p) => p.identifier?.toLowerCase().trim())
-      .filter(Boolean) as string[]
-
-    // 1. Allowlist: if onlyProcessRecipients is set and no TO matches → ignore
-    const hasAllowlist = settings.onlyProcessRecipients && settings.onlyProcessRecipients.length > 0
-    if (hasAllowlist) {
-      const anyToMatches = toEmails.some((e) =>
-        matchesFilterEntry(e, settings.onlyProcessRecipients!)
-      )
-      if (!anyToMatches) return true
-    }
-
-    // 2. Sender blocklist — always applies
-    if (settings.excludeSenders?.length && fromEmail) {
-      if (matchesFilterEntry(fromEmail, settings.excludeSenders)) return true
-    }
-
-    // 3. Recipient blocklist — skipped if allowlist was set (already validated TO)
-    if (!hasAllowlist && settings.excludeRecipients?.length) {
-      const allRecipients = [...toEmails, ...ccEmails]
-      if (allRecipients.some((e) => matchesFilterEntry(e, settings.excludeRecipients!))) return true
-    }
-
-    return false
-  }
-
-  /**
-   * Store minimal thread + message for dedup only. Skips participants, contacts, body, attachments.
-   */
-  private async storeIgnoredMessage(
-    messageData: MessageData
-  ): Promise<{ messageId: string; isNew: boolean }> {
-    const [thread] = await db
-      .insert(schema.Thread)
-      .values({
-        externalId: messageData.externalThreadId,
-        integrationId: messageData.integrationId,
-        organizationId: messageData.organizationId,
-        subject: messageData.subject ?? 'No Subject',
-        status: ThreadStatus.IGNORED,
-        firstMessageAt: messageData.sentAt,
-        lastMessageAt: messageData.sentAt,
-        messageCount: 1,
-        participantCount: 0,
-      })
-      .onConflictDoUpdate({
-        target: [schema.Thread.integrationId, schema.Thread.externalId],
-        set: {}, // Don't update existing threads — preserve their current status
-      })
-      .returning({ id: schema.Thread.id })
-
-    const [message] = await db
-      .insert(schema.Message)
-      .values({
-        externalId: messageData.externalId,
-        externalThreadId: messageData.externalThreadId,
-        threadId: thread.id,
-        organizationId: messageData.organizationId,
-        integrationId: messageData.integrationId,
-        internetMessageId: messageData.internetMessageId,
-        sentAt: messageData.sentAt,
-        receivedAt: messageData.receivedAt,
-        subject: messageData.subject ?? '',
-        isInbound: messageData.isInbound,
-      })
-      .onConflictDoNothing({
-        target: [schema.Message.integrationId, schema.Message.externalId],
-      })
-      .returning({ id: schema.Message.id })
-
-    logger.info('Stored ignored message (minimal, no body/contacts)', {
-      messageId: message?.id,
-      externalId: messageData.externalId,
-      threadId: thread.id,
-    })
-
-    return { messageId: message?.id ?? '', isNew: !!message }
-  }
-
-  /** Resolves the system user ID for an organization, with caching. */
-  private async resolveSystemUserId(organizationId: string): Promise<string> {
-    const cached = this.systemUserIdCache.get(organizationId)
-    if (cached) return cached
-
-    const systemUserId = await SystemUserService.getSystemUserForActions(organizationId)
-    this.systemUserIdCache.set(organizationId, systemUserId)
-    return systemUserId
-  }
-
-  /** Returns a cached UnifiedCrudHandler for the given organization. */
-  private async getCrudHandler(organizationId: string): Promise<UnifiedCrudHandler> {
-    let handler = this.crudHandlerCache.get(organizationId)
-    if (!handler) {
-      const systemUserId = await this.resolveSystemUserId(organizationId)
-      handler = new UnifiedCrudHandler(organizationId, systemUserId)
-      this.crudHandlerCache.set(organizationId, handler)
-    }
-    return handler
-  }
-
-  /** Returns the cached own-domains set for an org, fetching once per batch. */
-  private async resolveOwnDomains(organizationId: string): Promise<Set<string>> {
-    let ownDomains = this.ownDomainsByOrg.get(organizationId)
-    if (!ownDomains) {
-      ownDomains = await getOwnDomains(organizationId)
-      this.ownDomainsByOrg.set(organizationId, ownDomains)
-    }
-    return ownDomains
-  }
-
-  // ========================================================================
-  // Helper Functions (Internal)
-  // ========================================================================
-
-  /** Calculates initials (max 2) from a name string. */
-  private calculateInitials(name?: string | null): string | undefined {
-    if (!name) return undefined
-    return (
-      name
-        .trim()
-        .split(/\s+/)
-        .map((word) => word.charAt(0))
-        .filter((char) => char.match(/[a-zA-Z]/))
-        .slice(0, 2)
-        .join('')
-        .toUpperCase() || undefined
-    )
-  }
-
-  /** Calculates a display name, preferring name over identifier. */
-  private calculateDisplayName(
-    name?: string | null,
-    identifier?: string | null
-  ): string | undefined {
-    const trimmedName = name?.trim()
-    if (trimmedName) return trimmedName
-    const trimmedIdentifier = identifier?.trim()
-    if (trimmedIdentifier) {
-      if (trimmedIdentifier.includes('@')) return trimmedIdentifier
-      if (trimmedIdentifier.match(/^\+?\d+$/)) return trimmedIdentifier
-      if (trimmedIdentifier.length > 20) return trimmedIdentifier.substring(0, 15) + '...'
-      return trimmedIdentifier
-    }
-    return undefined
-  }
-
-  /** Normalizes an identifier based on its type. */
-  private normalizeIdentifier(identifier: string, type: IdentifierType): string {
-    const trimmed = identifier.trim()
-    switch (type) {
-      case IdentifierTypeEnum.EMAIL:
-        return trimmed.toLowerCase()
-      case IdentifierTypeEnum.PHONE: {
-        const digits = trimmed.replace(/\D/g, '')
-        return trimmed.startsWith('+') ? `+${digits}` : digits
-      }
-      default:
-        return trimmed
-    }
-  }
-
-  /** Determines IdentifierType based on integration provider and format guess. */
-  private async determineIdentifierType(
-    identifier: string,
-    integrationId: string
-  ): Promise<IdentifierType> {
-    // Get provider from integration
-    const [integration] = await db
-      .select({ provider: schema.Integration.provider })
-      .from(schema.Integration)
-      .where(and(eq(schema.Integration.id, integrationId), isNull(schema.Integration.deletedAt)))
-      .limit(1)
-
-    const provider = integration?.provider
-
-    switch (provider) {
-      case 'google':
-      case 'outlook':
-      case 'mailgun':
-      case 'email':
-        return 'EMAIL' as any
-      case 'openphone':
-      case 'sms':
-        return 'PHONE' as any
-      case 'facebook':
-        return 'FACEBOOK_PSID' as any
-      case 'instagram':
-        return 'INSTAGRAM_IGSID' as any
-      default: // Fallback guess
-        if (identifier.includes('@')) return 'EMAIL' as any
-        if (identifier.match(/^\+?\d{7,}$/)) return 'PHONE' as any // Basic phone guess
-        logger.warn(
-          `Could not reliably determine identifier type for integration provider ${provider}, identifier: ${identifier}. Defaulting to EMAIL.`
-        )
-        return 'EMAIL' as any // Risky fallback
-    }
-  }
-
-  /** Gets names from participant data. */
-  private getNamesFromParticipant(p: { name?: string | null; displayName?: string | null }): {
-    firstName?: string | null
-    lastName?: string | null
-  } {
-    const name = p.name?.trim()
-    if (!name) return { firstName: p.displayName, lastName: null }
-    const parts = name.split(' ').filter(Boolean)
-    if (parts.length <= 1) return { firstName: parts[0] || p.displayName, lastName: null }
-    return { firstName: parts.slice(0, -1).join(' '), lastName: parts[parts.length - 1] }
-  }
-
-  /**
-   * Extract Internet Message-ID from message data
-   */
-  private extractInternetMessageId(messageData: MessageData): string | undefined {
-    // Direct field
-    if (messageData.internetMessageId) {
-      return messageData.internetMessageId
-    }
-
-    // From headers in metadata
-    const metadata = messageData.metadata as any
-    const headers = metadata?.headers
-    if (headers?.['message-id']) {
-      return headers['message-id']
-    }
-    if (headers?.['Message-ID']) {
-      return headers['Message-ID']
-    }
-
-    // From internet headers array
-    if (messageData.internetHeaders && Array.isArray(messageData.internetHeaders)) {
-      for (const header of messageData.internetHeaders) {
-        const h = header as any
-        if (h.name?.toLowerCase() === 'message-id' && h.value) {
-          return h.value
-        }
-      }
-    }
-
-    return undefined
-  }
-
-  /**
-   * Check if two subjects are similar enough to be the same message
-   */
-  private isSimilarSubject(
-    subject1: string | null | undefined,
-    subject2: string | null | undefined
-  ): boolean {
-    if (!subject1 || !subject2) return false
-
-    // Normalize subjects
-    const normalize = (s: string) =>
-      s
-        .toLowerCase()
-        .replace(/^(re:|fwd:|fw:)\s*/gi, '')
-        .trim()
-    const normalized1 = normalize(subject1)
-    const normalized2 = normalize(subject2)
-
-    // Exact match after normalization
-    if (normalized1 === normalized2) return true
-
-    // Check if one contains the other (for truncated subjects)
-    if (normalized1.includes(normalized2) || normalized2.includes(normalized1)) return true
-
-    // Could add more sophisticated similarity checks (Levenshtein distance, etc.)
-    return false
-  }
-
-  /**
-   * Merge provider data into existing message without overwriting user content
-   */
-  private async mergeProviderData(existing: Message, providerData: MessageData): Promise<Message> {
-    // Don't overwrite user-authored fields
-    const mergedMetadata = {
-      ...((existing.metadata as any) || {}),
-      ...((providerData.metadata as any) || {}),
-      reconciled: true,
-      reconciledAt: new Date().toISOString(),
-    } as JsonValue
-
-    const [row] = await db
-      .update(schema.Message)
-      .set({
-        textPlain: existing.textPlain ?? providerData.textPlain,
-        textHtml: existing.textHtml ?? providerData.textHtml,
-        snippet: existing.snippet ?? providerData.snippet,
-        externalId: providerData.externalId ?? existing.externalId,
-        externalThreadId: providerData.externalThreadId ?? existing.externalThreadId,
-        hasAttachments: providerData.hasAttachments ?? existing.hasAttachments,
-        sendStatus: existing.sendStatus === 'PENDING' ? 'SENT' : existing.sendStatus,
-        metadata: mergedMetadata,
-        // lastModifiedTime: new Date(),
-      })
-      .where(eq(schema.Message.id, existing.id))
-      .returning()
-
-    return row
-  }
-
-  /**
-   * Reconcile message with existing data to prevent duplicates
-   * Delegates to the MessageReconcilerService when available
-   */
-  private async reconcileMessage(messageData: MessageData): Promise<Message | null> {
-    // Use the new reconciler service if available
-    if (this.reconciler && this.organizationId) {
-      const result = await this.reconciler.reconcileIncomingSync(messageData)
-
-      if (result.isReconciled && result.existingMessageId) {
-        // Get the full message object to return
-        const [message] = await db
-          .select()
-          .from(schema.Message)
-          .where(eq(schema.Message.id, result.existingMessageId))
-          .limit(1)
-
-        if (message) {
-          logger.debug('Message reconciled by MessageReconcilerService', {
-            messageId: message.id,
-            externalId: messageData.externalId,
-          })
-          return message
-        }
-      }
-
-      // No reconciliation found
-      return null
-    }
-
-    // Fallback to legacy reconciliation if reconciler not available
-    // Priority 1: Provider ID match
-    if (messageData.externalId && messageData.integrationId) {
-      try {
-        const [existing] = await db
-          .select()
-          .from(schema.Message)
-          .where(
-            and(
-              eq(schema.Message.integrationId, messageData.integrationId),
-              eq(schema.Message.externalId, messageData.externalId)
-            )
-          )
-          .limit(1)
-        if (existing) {
-          logger.debug('Message reconciled by externalId (legacy)', {
-            messageId: existing.id,
-            externalId: messageData.externalId,
-          })
-          return await this.mergeProviderData(existing, messageData)
-        }
-      } catch (error) {
-        logger.error('Error checking externalId match', { error })
-      }
-    }
-
-    // Priority 2: Internet Message-ID match
-    const internetMessageId = this.extractInternetMessageId(messageData)
-    if (internetMessageId && messageData.organizationId) {
-      try {
-        const [existing] = await db
-          .select()
-          .from(schema.Message)
-          .where(
-            and(
-              eq(schema.Message.organizationId, messageData.organizationId),
-              eq(schema.Message.internetMessageId, internetMessageId)
-            )
-          )
-          .limit(1)
-        if (existing) {
-          logger.debug('Message reconciled by internetMessageId (legacy)', {
-            messageId: existing.id,
-            internetMessageId,
-          })
-          return await this.mergeProviderData(existing, messageData)
-        }
-      } catch (error) {
-        logger.error('Error checking internetMessageId match', { error })
-      }
-    }
-
-    // Priority 3: Heuristic match (careful!)
-    // Only attempt if we have enough data
-    if (messageData.externalThreadId && messageData.sentAt && messageData.from?.identifier) {
-      try {
-        // First need to find the thread
-        const [thread] = await db
-          .select()
-          .from(schema.Thread)
-          .where(
-            and(
-              eq(schema.Thread.integrationId, messageData.integrationId),
-              eq(schema.Thread.externalId, messageData.externalThreadId)
-            )
-          )
-          .limit(1)
-
-        if (thread) {
-          const timeWindowStart = new Date(messageData.sentAt.getTime() - 2 * 60 * 1000) // 2 minutes before
-          const timeWindowEnd = new Date(messageData.sentAt.getTime() + 2 * 60 * 1000) // 2 minutes after
-
-          const candidates = await db.query.Message.findMany({
-            where: (t, { and, eq, gte, lte }) =>
-              and(
-                eq(t.threadId, thread.id),
-                gte(t.sentAt, timeWindowStart as any),
-                lte(t.sentAt, timeWindowEnd as any)
-              ),
-            with: { from: true },
-          })
-
-          // Check for matching sender and similar subject
-          const match = candidates.find((c) => {
-            const senderMatch = c.from?.identifier === messageData.from.identifier
-            const subjectMatch = this.isSimilarSubject(c.subject, messageData.subject)
-            return senderMatch && subjectMatch
-          })
-
-          if (match) {
-            logger.debug('Message reconciled by heuristic match (legacy)', {
-              messageId: match.id,
-              threadId: thread.id,
-              subject: messageData.subject,
-            })
-            return await this.mergeProviderData(match, messageData)
-          }
-        }
-      } catch (error) {
-        logger.error('Error checking heuristic match', { error })
-      }
-    }
-
-    // No match found
-    return null
-  }
-
-  /**
-   * Sets the integration settings for this storage service instance.
-   * @param settings The integration settings to use
-   */
-  setIntegrationSettings(settings: IntegrationSettings | undefined) {
+  setIntegrationSettings(settings: IntegrationSettings | undefined): void {
     this.integrationSettings = settings
-  }
-
-  /**
-   * Checks if the organization has sent a message to a participant.
-   * @param participantId The ID of the participant
-   * @param identifier The participant's identifier (email/phone)
-   * @param organizationId The ID of the organization
-   * @returns Promise<boolean> True if organization has sent to this participant
-   */
-  private async hasOrganizationSentToParticipant(
-    participantId: string,
-    identifier: string,
-    organizationId: string
-  ): Promise<boolean> {
-    // Check cache first (persists across batches during initial sync)
-    const cachedResult = await this.selectiveCache.hasSentToRecipient(identifier, organizationId)
-    if (cachedResult) {
-      return true
-    }
-
-    // Check the participant's hasReceivedMessage flag in database
-    const [participant] = await db
-      .select({
-        hasReceivedMessage: schema.Participant.hasReceivedMessage,
+    for (const ctxPromise of this.ctxByOrg.values()) {
+      ctxPromise.then((ctx) => {
+        ctx.integrationSettings = settings
       })
-      .from(schema.Participant)
-      .where(eq(schema.Participant.id, participantId))
-      .limit(1)
-
-    if (participant?.hasReceivedMessage) {
-      // Cache this result for future checks
-      await this.selectiveCache.markSentToRecipient(identifier, organizationId)
-      return true
-    }
-
-    // Fallback: Query message history
-    const [sentMessage] = await db
-      .select({
-        id: schema.MessageParticipant.id,
-      })
-      .from(schema.MessageParticipant)
-      .innerJoin(schema.Message, eq(schema.MessageParticipant.messageId, schema.Message.id))
-      .where(
-        and(
-          eq(schema.MessageParticipant.participantId, participantId),
-          inArray(schema.MessageParticipant.role, [
-            ParticipantRoleEnum.TO,
-            ParticipantRoleEnum.CC,
-            ParticipantRoleEnum.BCC,
-          ]),
-          eq(schema.Message.organizationId, organizationId),
-          eq(schema.Message.isInbound, false)
-        )
-      )
-      .limit(1)
-
-    if (sentMessage) {
-      // Update the participant record for future reference
-      await db
-        .update(schema.Participant)
-        .set({
-          hasReceivedMessage: true,
-          lastSentMessageAt: new Date(),
-          updatedAt: new Date(),
-        })
-        .where(eq(schema.Participant.id, participantId))
-      // Cache this result
-      await this.selectiveCache.markSentToRecipient(identifier, organizationId)
-      return true
-    }
-
-    return false
-  }
-
-  /**
-   * Finds or creates a Contact (EntityInstance) for a Participant.
-   * Uses UnifiedCrudHandler to query/create via EntityInstance + FieldValue.
-   * Respects integration record creation mode (none/selective/all).
-   */
-  private async findOrCreateContactForParticipant(
-    participant: Participant,
-    organizationId: string,
-    messageContext?: { isInbound: boolean; role: ParticipantRole }
-  ): Promise<string | null> {
-    try {
-      const mode = this.integrationSettings?.recordCreation?.mode || 'selective'
-      const handler = await this.getCrudHandler(organizationId)
-
-      // Determine which system attribute to search by
-      const systemAttr =
-        participant.identifierType === IdentifierTypeEnum.PHONE ? 'phone' : 'primary_email'
-
-      // Mode: NONE — lookup only, never create
-      if (mode === 'none') {
-        const existing = await handler.findByField('contact', systemAttr, participant.identifier)
-        return existing?.id ?? null
-      }
-
-      // Mode: SELECTIVE — only create for outbound recipients or previously-contacted participants
-      if (mode === 'selective' && messageContext) {
-        const existing = await handler.findByField('contact', systemAttr, participant.identifier)
-        if (existing) return existing.id
-
-        const isOutboundRecipient =
-          !messageContext.isInbound &&
-          [ParticipantRoleEnum.TO, ParticipantRoleEnum.CC, ParticipantRoleEnum.BCC].includes(
-            messageContext.role
-          )
-
-        if (!isOutboundRecipient) {
-          const hasSentBefore = await this.hasOrganizationSentToParticipant(
-            participant.id,
-            participant.identifier,
-            organizationId
-          )
-          if (!hasSentBefore) {
-            logger.info(
-              `Skipping contact creation for inbound-only participant ${participant.id} (selective mode)`
-            )
-            return null
-          }
-        }
-      }
-
-      // Mode: ALL (or selective criteria met) — find or create
-      const names = this.getNamesFromParticipant(participant)
-      const findBy: Record<string, unknown> = { [systemAttr]: participant.identifier }
-      const createValues: Record<string, unknown> = {
-        first_name: names.firstName,
-        last_name: names.lastName,
-        contact_status: 'ACTIVE',
-      }
-
-      // For phone contacts, ensure phone is in create values too
-      if (participant.identifierType === IdentifierTypeEnum.PHONE) {
-        createValues.phone = participant.identifier
-      }
-
-      const { instance } = await handler.findOrCreate('contact', findBy, createValues)
-      const contactId = instance.id
-
-      // Auto-link contact to a company based on its email domain.
-      // Swallows errors — contact creation must succeed even if linking fails.
-      if (contactId) {
-        const ownDomains = await this.resolveOwnDomains(organizationId)
-        await linkContactToCompanyByDomain({
-          organizationId,
-          crudHandler: handler,
-          contactId,
-          identifier: participant.identifier,
-          identifierType: participant.identifierType,
-          companyIdByDomain: this.companyIdByDomain,
-          ownDomains,
-        })
-      }
-
-      return contactId
-    } catch (error) {
-      logger.error('Error finding/creating contact for participant:', {
-        error,
-        participantId: participant.id,
-      })
-      throw error
     }
   }
 
-  /**
-   * Finds or creates a Participant record, ensuring it's linked to a Contact.
-   */
-  private async findOrCreateParticipantRecord(
-    participantInput: ParticipantInputData,
-    identifierType: IdentifierType,
-    organizationId: string,
-    messageContext?: { isInbound: boolean; role: ParticipantRole }
-  ): Promise<Participant> {
-    if (!participantInput.identifier) {
-      throw new Error('Participant identifier cannot be empty.')
-    }
-    const normalizedIdentifier = this.normalizeIdentifier(
-      participantInput.identifier,
-      identifierType
-    )
-    const name = participantInput.name?.trim() || null
-
-    try {
-      // Use Drizzle's upsert for Participant
-      const initials = this.calculateInitials(name)
-      const displayName = this.calculateDisplayName(name, normalizedIdentifier)
-
-      // Determine if this is the first interaction and track it
-      const isOutboundRecipient =
-        messageContext &&
-        !messageContext.isInbound &&
-        [ParticipantRoleEnum.TO, ParticipantRoleEnum.CC, ParticipantRoleEnum.BCC].includes(
-          messageContext.role
-        )
-
-      const participantData = await db
-        .insert(schema.Participant)
-        .values({
-          identifier: normalizedIdentifier,
-          identifierType: identifierType,
-          name,
-          displayName,
-          initials,
-          organizationId,
-          // Set tracking fields on creation
-          ...(messageContext && {
-            firstInteractionType: messageContext.isInbound ? 'received' : 'sent',
-            firstInteractionDate: new Date(),
-            hasReceivedMessage: isOutboundRecipient || false,
-            lastSentMessageAt: isOutboundRecipient ? new Date() : null,
-          }),
-          updatedAt: new Date(),
-        })
-        .onConflictDoUpdate({
-          target: [
-            schema.Participant.organizationId,
-            schema.Participant.identifier,
-            schema.Participant.identifierType,
-          ],
-          set: {
-            // Update name/display info if participant already exists
-            ...(name !== undefined && { name: name }),
-            ...(displayName !== undefined && { displayName: displayName }),
-            ...(initials !== undefined && { initials: initials }),
-            updatedAt: new Date(),
-            // Update tracking fields if we're sending to them
-            ...(isOutboundRecipient && {
-              hasReceivedMessage: true,
-              lastSentMessageAt: new Date(),
-            }),
-          },
-        })
-        .returning()
-
-      const participant = participantData[0]
-
-      // Ensure linked to contact (if appropriate based on settings)
-      if (!participant.entityInstanceId) {
-        const entityInstanceId = await this.findOrCreateContactForParticipant(
-          participant,
-          organizationId,
-          messageContext
-        )
-        // Only update participant to link contact if a contact was created/found
-        if (entityInstanceId) {
-          const updatedParticipants = await db
-            .update(schema.Participant)
-            .set({ entityInstanceId, updatedAt: new Date() })
-            .where(eq(schema.Participant.id, participant.id))
-            .returning()
-
-          return updatedParticipants[0]
-        }
-      }
-
-      return participant
-    } catch (error) {
-      logger.error('Error upserting participant record:', {
-        error,
-        identifier: normalizedIdentifier,
-        type: identifierType,
-      })
-      throw error
-    }
-  }
-
-  /** Store an attachment */
-  /**
-   * Efficiently updates thread metadata using aggregate SQL
-   * This runs a single query instead of multiple select + update queries
-   */
-  private async updateThreadMetadataEfficient(threadId: string): Promise<void> {
-    try {
-      await db.execute(sql`
-        UPDATE "Thread" t
-        SET
-          "messageCount" = COALESCE((
-            SELECT COUNT(*)
-            FROM "Message"
-            WHERE "threadId" = ${threadId}
-              AND "sentAt" IS NOT NULL
-          ), 0),
-          "firstMessageAt" = (
-            SELECT MIN("sentAt")
-            FROM "Message"
-            WHERE "threadId" = ${threadId}
-              AND "sentAt" IS NOT NULL
-          ),
-          "lastMessageAt" = (
-            SELECT MAX("sentAt")
-            FROM "Message"
-            WHERE "threadId" = ${threadId}
-              AND "sentAt" IS NOT NULL
-          ),
-          "latestMessageId" = (
-            SELECT id
-            FROM "Message"
-            WHERE "threadId" = ${threadId}
-            ORDER BY "receivedAt" DESC NULLS LAST,
-                     "sentAt" DESC NULLS LAST,
-                     id DESC
-            LIMIT 1
-          ),
-          "participantCount" = COALESCE((
-            SELECT COUNT(DISTINCT "participantId")
-            FROM "MessageParticipant" mp
-            JOIN "Message" m ON mp."messageId" = m.id
-            WHERE m."threadId" = ${threadId}
-              AND mp."participantId" IS NOT NULL
-          ), 0)
-        WHERE t.id = ${threadId}
-      `)
-
-      logger.debug('Efficiently updated thread metadata', { threadId })
-    } catch (error) {
-      logger.error('Failed to update thread metadata efficiently', { threadId, error })
-      // Don't throw - this is a non-critical operation
-    }
-  }
-
-  // ========================================================================
-  // Core Public Methods
-  // ========================================================================
-
-  /**
-   * Stores a message and links its participants using the REVISED schema (Message.fromId -> Participant.id).
-   * @param messageData - The message data object to store
-   * @returns The stored message ID and whether it was a new insert (vs reconciled with existing).
-   */
-  async storeMessage(messageData: MessageData): Promise<{ messageId: string; isNew: boolean }> {
-    // Input validation
-    if (!messageData.from?.identifier) {
-      throw new Error(
-        `Message externalId ${messageData.externalId} is missing required sender identifier.`
-      )
-    }
-
-    try {
-      // NEW: Reconcile by Message-ID first (prevents duplicate threads/messages)
-      const existingByMsgId = messageData.internetMessageId
-        ? (
-            await db
-              .select({
-                id: schema.Message.id,
-                threadId: schema.Message.threadId,
-                externalId: schema.Message.externalId,
-                textPlain: schema.Message.textPlain,
-                textHtml: schema.Message.textHtml,
-              })
-              .from(schema.Message)
-              .where(
-                and(
-                  eq(schema.Message.organizationId, messageData.organizationId),
-                  eq(schema.Message.internetMessageId, messageData.internetMessageId)
-                )
-              )
-              .limit(1)
-          )?.[0]
-        : null
-
-      if (existingByMsgId) {
-        logger.info('Reconciling message by internetMessageId', {
-          messageId: existingByMsgId.id,
-          internetMessageId: messageData.internetMessageId,
-          incomingExternalId: messageData.externalId,
-        })
-
-        // Update message with provider data (don't overwrite user content)
-        await db
-          .update(schema.Message)
-          .set({
-            externalId: messageData.externalId,
-            externalThreadId: messageData.externalThreadId,
-            textPlain: existingByMsgId.textPlain ?? messageData.textPlain,
-            textHtml: messageData.htmlBodyStorageLocationId
-              ? null
-              : (existingByMsgId.textHtml ?? messageData.textHtml),
-            snippet: messageData.snippet ?? null,
-            htmlBodyStorageLocationId: messageData.htmlBodyStorageLocationId ?? undefined,
-            hasAttachments: messageData.hasAttachments,
-            metadata: messageData.metadata ?? {},
-            receivedAt: messageData.receivedAt,
-            sentAt: messageData.sentAt,
-            historyId: messageData.historyId ? BigInt(messageData.historyId) : null,
-            isInbound: messageData.isInbound,
-            updatedAt: new Date(),
-          })
-          .where(eq(schema.Message.id, existingByMsgId.id))
-
-        // Update thread's externalId if it's still a placeholder
-        if (messageData.externalThreadId && existingByMsgId.threadId) {
-          const [thread] = await db
-            .select({ externalId: schema.Thread.externalId })
-            .from(schema.Thread)
-            .where(eq(schema.Thread.id, existingByMsgId.threadId))
-            .limit(1)
-
-          if (
-            thread &&
-            (!thread.externalId ||
-              thread.externalId.startsWith('new_') ||
-              thread.externalId.startsWith('pending_'))
-          ) {
-            await db
-              .update(schema.Thread)
-              .set({ externalId: messageData.externalThreadId })
-              .where(eq(schema.Thread.id, existingByMsgId.threadId))
-
-            logger.info('Updated thread with real externalId', {
-              threadId: existingByMsgId.threadId,
-              externalId: messageData.externalThreadId,
-            })
-          }
-        }
-
-        // Update thread metadata
-        await this.updateThreadMetadataEfficient(existingByMsgId.threadId)
-
-        // Done: we reconciled, no thread upsert needed
-        return { messageId: existingByMsgId.id, isNew: false }
-      }
-
-      // CRITICAL: Always check reconciliation with MessageReconcilerService first
-      // This prevents duplicate threads by matching sent messages during sync
-      const existingMessage = await this.reconcileMessage(messageData)
-      if (existingMessage) {
-        logger.info('Message reconciled with existing record via MessageReconcilerService', {
-          messageId: existingMessage.id,
-          externalId: messageData.externalId,
-        })
-
-        // Promote thread externalId if needed
-        if (existingMessage.threadId && messageData.externalThreadId) {
-          const [thread] = await db
-            .select({ externalId: schema.Thread.externalId })
-            .from(schema.Thread)
-            .where(eq(schema.Thread.id, existingMessage.threadId))
-            .limit(1)
-
-          const ext = thread?.externalId
-          if (
-            !ext ||
-            ext.startsWith('new_') ||
-            ext.startsWith('pending_') ||
-            ext.startsWith('draft_') ||
-            (ext.includes('-') && ext.length === 36)
-          ) {
-            await db
-              .update(schema.Thread)
-              .set({ externalId: messageData.externalThreadId })
-              .where(eq(schema.Thread.id, existingMessage.threadId))
-
-            logger.info('Promoted thread externalId from placeholder during reconciliation', {
-              threadId: existingMessage.threadId,
-              oldExternalId: ext,
-              newExternalId: messageData.externalThreadId,
-            })
-          }
-        }
-
-        // Update thread metadata if needed
-        if (existingMessage.threadId) {
-          await this.updateThreadMetadataEfficient(existingMessage.threadId)
-        }
-
-        return { messageId: existingMessage.id, isNew: false }
-      }
-
-      logger.info('Storing new message (Schema: Msg->Participant)', {
-        externalId: messageData.externalId,
-        integrationId: messageData.integrationId,
-      })
-
-      // Fetch integration settings if not already loaded
-      if (!this.integrationSettings && messageData.integrationId) {
-        const [integration] = await db
-          .select({ metadata: schema.Integration.metadata })
-          .from(schema.Integration)
-          .where(
-            and(
-              eq(schema.Integration.id, messageData.integrationId),
-              isNull(schema.Integration.deletedAt)
-            )
-          )
-          .limit(1)
-
-        this.integrationSettings = (integration?.metadata as any)?.settings as
-          | IntegrationSettings
-          | undefined
-      }
-
-      // Check email filtering rules — early exit with minimal storage for dedup
-      if (this.shouldIgnoreMessage(messageData)) {
-        return this.storeIgnoredMessage(messageData)
-      }
-
-      // --- 1. Process ALL Participants & Cache Results ---
-      const participantCache = new Map<string, Participant>() // Key: "TYPE:normalized_identifier"
-      const participantInputsWithRoles: Array<{
-        role: ParticipantRole
-        data: ParticipantInputData
-      }> = []
-
-      const processAndCacheParticipant = async (
-        data: ParticipantInputData,
-        role?: ParticipantRole
-      ): Promise<Participant | null> => {
-        if (!data?.identifier) return null
-        const identifierType = await this.determineIdentifierType(
-          data.identifier,
-          messageData.integrationId
-        )
-        const normalizedId = this.normalizeIdentifier(data.identifier, identifierType)
-        const cacheKey = `${identifierType}:${normalizedId}`
-
-        if (participantCache.has(cacheKey)) {
-          return participantCache.get(cacheKey)!
-        } else {
-          // Pass message context when creating participants
-          const messageContext = role
-            ? {
-                isInbound: messageData.isInbound,
-                role: role,
-              }
-            : undefined
-
-          const participantRecord = await this.findOrCreateParticipantRecord(
-            data,
-            identifierType,
-            messageData.organizationId,
-            messageContext
-          )
-          participantCache.set(cacheKey, participantRecord)
-          return participantRecord
-        }
-      }
-
-      // Collect roles and ensure all unique participants are processed and cached
-      const allInputs = [
-        { role: ParticipantRoleEnum.FROM, data: messageData.from },
-        ...messageData.to.map((p) => ({ role: ParticipantRoleEnum.TO, data: p })),
-        ...(messageData.cc || []).map((p) => ({ role: ParticipantRoleEnum.CC, data: p })),
-        ...(messageData.bcc || []).map((p) => ({ role: ParticipantRoleEnum.BCC, data: p })),
-        ...(messageData.replyTo || []).map((p) => ({
-          role: ParticipantRoleEnum.REPLY_TO,
-          data: p,
-        })),
-      ]
-
-      for (const { role, data } of allInputs) {
-        if (data?.identifier) {
-          participantInputsWithRoles.push({ role, data }) // Keep track of all roles
-          await processAndCacheParticipant(data, role) // Ensure participant exists in cache with role context
-        } else {
-          logger.warn('Skipping participant input due to missing identifier', { role })
-        }
-      }
-
-      // --- 2. Get Sender and ReplyTo Participant IDs from Cache ---
-      const senderParticipant = await processAndCacheParticipant(
-        messageData.from,
-        ParticipantRoleEnum.FROM
-      ) // Ensures sender is processed
-      if (!senderParticipant) {
-        throw new Error(
-          `Failed to process sender participant for message ${messageData.externalId}`
-        )
-      }
-      const senderParticipantId = senderParticipant.id
-
-      let firstReplyToParticipantId: string | null = null
-      if (messageData.replyTo && messageData.replyTo.length > 0) {
-        const replyToParticipant = await processAndCacheParticipant(
-          messageData.replyTo[0],
-          ParticipantRoleEnum.REPLY_TO
-        )
-        firstReplyToParticipantId = replyToParticipant?.id ?? null
-      }
-
-      // --- 3. Prepare participant IDs for thread ---
-      // Extract all unique participant IDs from the current message
-      const currentMessageParticipantIds: string[] = []
-      for (const participant of participantCache.values()) {
-        if (participant?.id) {
-          currentMessageParticipantIds.push(participant.id)
-        }
-      }
-
-      // --- 4. Upsert Thread ---
-      const threadData = await db
-        .insert(schema.Thread)
-        .values({
-          externalId: messageData.externalThreadId,
-          integrationId: messageData.integrationId,
-          organizationId: messageData.organizationId,
-          inboxId: messageData.inboxId ?? null,
-          subject: messageData.subject ?? 'No Subject',
-          status: ThreadStatus.OPEN,
-          firstMessageAt: messageData.sentAt,
-          lastMessageAt: messageData.sentAt,
-          messageCount: 1, // This is the first message
-          participantCount: currentMessageParticipantIds.length,
-        })
-        .onConflictDoUpdate({
-          target: [schema.Thread.integrationId, schema.Thread.externalId],
-          set: {
-            // Minimal updates for existing threads
-            subject: messageData.subject || undefined,
-            inboxId: messageData.inboxId ?? undefined,
-          },
-        })
-        .returning({
-          id: schema.Thread.id,
-          messageCount: schema.Thread.messageCount,
-          firstMessageAt: schema.Thread.firstMessageAt,
-          lastMessageAt: schema.Thread.lastMessageAt,
-          participantCount: schema.Thread.participantCount,
-        })
-
-      const thread = threadData[0]
-      const isNewThread =
-        (thread.messageCount ?? 0) === 1 &&
-        thread.firstMessageAt?.getTime() === messageData.sentAt.getTime()
-
-      // --- 4. Upsert Message ---
-      const messageRecords = await db
-        .insert(schema.Message)
-        .values({
-          externalThreadId: messageData.externalThreadId,
-          threadId: thread.id,
-          organizationId: messageData.organizationId,
-          integrationId: messageData.integrationId,
-          historyId: messageData.historyId ? Number(messageData.historyId) : null,
-          createdAt: messageData.createdTime,
-          updatedAt: new Date(),
-          sentAt: messageData.sentAt,
-          receivedAt: messageData.receivedAt,
-          internetMessageId:
-            this.extractInternetMessageId(messageData) || messageData.internetMessageId,
-          subject: messageData.subject ?? '',
-          hasAttachments: messageData.hasAttachments,
-          textHtml: messageData.htmlBodyStorageLocationId ? null : messageData.textHtml,
-          textPlain: messageData.textPlain,
-          snippet: messageData.snippet,
-          htmlBodyStorageLocationId: messageData.htmlBodyStorageLocationId ?? null,
-          metadata: messageData.metadata || null,
-          isInbound: messageData.isInbound,
-          isFirstInThread: isNewThread,
-
-          fromId: senderParticipantId, // Set direct link on create
-          replyToId: firstReplyToParticipantId, // Set direct link on create (optional)
-        })
-        .onConflictDoUpdate({
-          target: [schema.Message.integrationId, schema.Message.externalId],
-          set: {
-            threadId: thread.id,
-            historyId: messageData.historyId ? Number(messageData.historyId) : null,
-            updatedAt: new Date(),
-            sentAt: messageData.sentAt,
-            receivedAt: messageData.receivedAt,
-            subject: messageData.subject || '',
-            hasAttachments: messageData.hasAttachments,
-            textHtml: messageData.htmlBodyStorageLocationId ? null : messageData.textHtml,
-            textPlain: messageData.textPlain,
-            snippet: messageData.snippet,
-            htmlBodyStorageLocationId: messageData.htmlBodyStorageLocationId ?? null,
-            metadata: messageData.metadata || null,
-            isInbound: messageData.isInbound,
-            fromId: senderParticipantId, // Update direct link
-            replyToId: firstReplyToParticipantId, // Update direct link
-          },
-        })
-        .returning({ id: schema.Message.id })
-
-      const messageRecord = messageRecords[0]
-
-      // --- 4b. Set latestMessageId for new threads ---
-      // Cannot be set during thread creation because the message doesn't exist yet
-      if (isNewThread && messageRecord?.id) {
-        await db
-          .update(schema.Thread)
-          .set({ latestMessageId: messageRecord.id })
-          .where(eq(schema.Thread.id, thread.id))
-      }
-
-      // --- 5. Create MessageParticipant Links ---
-      const messageParticipantData: any[] = []
-      for (const { role, data } of participantInputsWithRoles) {
-        if (!data?.identifier) continue
-        const identifierType = await this.determineIdentifierType(
-          data.identifier,
-          messageData.integrationId
-        )
-        const normalizedId = this.normalizeIdentifier(data.identifier, identifierType)
-        const participantId = participantCache.get(`${identifierType}:${normalizedId}`)?.id
-        if (participantId) {
-          messageParticipantData.push({
-            messageId: messageRecord.id,
-            participantId: participantId,
-            role: role,
-          })
-        } else {
-          logger.error(
-            `Participant ID not found in cache for ${normalizedId} while creating MessageParticipant links.`
-          )
-        }
-      }
-      if (messageParticipantData.length > 0) {
-        // Use insert with onConflictDoNothing for idempotency
-        await db
-          .insert(schema.MessageParticipant)
-          .values(messageParticipantData)
-          .onConflictDoNothing()
-
-        logger.debug(
-          `Created/Skipped ${messageParticipantData.length} MessageParticipant links for message ${messageRecord.id}`
-        )
-      }
-
-      // --- 6. Store Attachments ---
-      // Attachment storage is now handled by the ingest pipeline via canonical Attachment + MediaAsset
-
-      // --- 7. Update Thread Metadata (if needed) ---
-      // Smart decision: only update if thread metadata might be stale
-      const shouldUpdateThreadMetadata =
-        !isNewThread &&
-        (!thread.firstMessageAt ||
-          !thread.lastMessageAt ||
-          messageData.sentAt < thread.firstMessageAt ||
-          messageData.sentAt > thread.lastMessageAt)
-
-      if (shouldUpdateThreadMetadata) {
-        // Use efficient aggregate update instead of multiple queries
-        await this.updateThreadMetadataEfficient(thread.id)
-      }
-
-      logger.info('Message stored successfully (Revised Schema v2)', {
-        messageId: messageRecord.id,
-        externalId: messageData.externalId,
-      })
-      return { messageId: messageRecord.id, isNew: true }
-    } catch (error: any) {
-      logger.error('Error storing message (Revised Schema v2):', {
-        error: error.message,
-        externalId: messageData?.externalId ?? 'UNKNOWN',
-        stack: error.stack,
-      })
-      // Check for unique constraint violation (P2002, similar concept in PostgreSQL)
-      if (error.message?.includes('duplicate key') || error.code === '23505') {
-        logger.warn(
-          `Unique constraint violation storing message ${messageData?.externalId ?? 'UNKNOWN'}. Assuming already processed.`
-        )
-        const [existing] = await db
-          .select({ id: schema.Message.id })
-          .from(schema.Message)
-          .where(
-            and(
-              eq(schema.Message.integrationId, messageData.integrationId),
-              eq(schema.Message.externalId, messageData.externalId)
-            )
-          )
-          .limit(1)
-
-        if (existing) return { messageId: existing.id, isNew: false }
-      }
-      throw error
-    }
-  }
-
-  /**
-   * Stores multiple messages in batch mode with optimizations for initial sync.
-   * @param messages Array of message data to store
-   * @param batchId Optional batch identifier for tracking
-   * @param isInitialSync Whether this is part of an initial sync operation
-   * @returns Number of successfully stored messages
-   */
-  async batchStoreMessages(
-    messages: MessageData[],
-    batchId?: string,
-    isInitialSync: boolean = false
-  ): Promise<number> {
-    if (messages.length === 0) return 0
-
-    const organizationId = messages[0]?.organizationId
-    if (!organizationId) {
-      logger.error('No organizationId found in batch messages')
-      return 0
-    }
-
-    // Generate batch ID if not provided
-    const actualBatchId = batchId || uuidv4()
-
-    // Track batch processing if this is initial sync
-    if (isInitialSync) {
-      this.isInitialSync = true
-      await this.selectiveCache.markBatchProcessing(organizationId, actualBatchId)
-    }
-
-    // Reset per-batch caches so they're scoped to this batch.
-    this.companyIdByDomain.clear()
-    this.ownDomainsByOrg.clear()
-
-    // Sort messages chronologically for accurate selective mode processing
-    const sortedMessages = [...messages].sort(
-      (a, b) => (a.sentAt?.getTime() || 0) - (b.sentAt?.getTime() || 0)
-    )
-
-    logger.info(`Starting batch store for ${messages.length} messages (sorted chronologically)`, {
-      batchId: actualBatchId,
-      isInitialSync,
-      organizationId,
-    })
-
-    let successCount = 0
-    for (const message of sortedMessages) {
-      try {
-        await this.storeMessage(message)
-        successCount++
-
-        // Update cache for outbound messages during initial sync
-        if (isInitialSync && !message.isInbound) {
-          const recipients = [...message.to, ...(message.cc || []), ...(message.bcc || [])]
-
-          const recipientIdentifiers = recipients
-            .map((r) => r?.identifier)
-            .filter(Boolean) as string[]
-
-          if (recipientIdentifiers.length > 0) {
-            await this.selectiveCache.markMultipleSentToRecipients(
-              recipientIdentifiers,
-              organizationId
-            )
-          }
-        }
-      } catch (error) {
-        logger.error('Error storing message in batch:', {
-          error: (error as Error).message,
-          externalId: message.externalId,
-          batchId: actualBatchId,
-        })
-        // Continue with the next message
-      }
-    }
-
-    // Complete batch tracking
-    if (isInitialSync) {
-      await this.selectiveCache.completeBatch(organizationId, actualBatchId, successCount)
-      this.isInitialSync = false
-    }
-
-    logger.info(`Batch store completed: ${successCount} of ${messages.length} messages stored.`, {
-      batchId: actualBatchId,
-      organizationId,
-    })
-
-    return successCount
-  }
-
-  /**
-   * Sets the initial sync mode for the storage service.
-   * @param enabled Whether initial sync mode is enabled
-   */
-  setInitialSyncMode(enabled: boolean) {
+  setInitialSyncMode(enabled: boolean): void {
     this.isInitialSync = enabled
-    if (!enabled) {
-      // Initial sync completed, cache will persist with TTL
-      logger.info('Initial sync mode disabled - cache will persist for subsequent batches')
-    }
-  }
-
-  // --- Retrieval Methods (Update Includes) ---
-
-  /** Retrieves messages for a specific thread, including participant details */
-  async getThreadMessages(threadId: string): Promise<Message[]> {
-    try {
-      const messages = await db.query.Message.findMany({
-        where: (messages, { eq }) => eq(messages.threadId, threadId),
-        orderBy: (messages, { asc }) => [asc(messages.sentAt)],
-        with: {
-          from: true, // Include the direct sender Participant
-          replyTo: true, // Include the direct replyTo Participant (optional)
-          participants: {
-            // Include the join table records
-            orderBy: (participants, { asc }) => [asc(participants.role)], // Consistent role order
-            with: {
-              participant: true, // Include the actual Participant details
-            },
-          },
-          attachments: true,
-        },
+    for (const ctxPromise of this.ctxByOrg.values()) {
+      ctxPromise.then((ctx) => {
+        ctx.isInitialSync = enabled
       })
-      return messages as Message[]
-    } catch (error) {
-      logger.error('Error getting thread messages:', { error, threadId })
-      throw error
     }
   }
 
-  /** Retrieves a specific thread by its internal ID */
-  async getThread(threadId: string, organizationId: string): Promise<Thread | null> {
-    // Added organizationId parameter for security/scoping
-    try {
-      const thread = await db.query.Thread.findFirst({
-        where: (threads, { and, eq }) =>
-          and(eq(threads.id, threadId), eq(threads.organizationId, organizationId)),
-        with: {
-          labels: { with: { label: true } },
-          // tags: TEMPORARILY DISABLED - migration to FieldValue in progress
-          // tags: { with: { tag: true } },
-          assignee: true,
-          // Note: inbox relation removed - Thread.inboxId was removed in migration 0028
-          integration: true, // Added to support provider-based type derivation
-          // Include first/last message snippet + sender participant?
-          messages: {
-            orderBy: (messages, { desc }) => [desc(messages.sentAt)], // Get last message
-            limit: 1,
-            with: { from: true }, // Include sender Participant
-          },
-        },
-      })
-      return thread as Thread | null
-    } catch (error) {
-      logger.error('Error getting thread:', { error, threadId })
-      throw error
-    }
+  async storeMessage(
+    messageData: Parameters<typeof storeMessage>[1]
+  ): Promise<{ messageId: string; isNew: boolean }> {
+    const ctx = await this.resolveCtx(messageData.organizationId)
+    return storeMessage(ctx, messageData)
   }
 
-  // getThreads, getAllProviderMessages etc. would need similar reviews of their 'include' and 'where' clauses
-  // based on how participant filtering/display is required.
+  async batchStoreMessages(
+    messages: Parameters<typeof batchStoreMessages>[1],
+    batchId?: string,
+    isInitialSync = false
+  ): Promise<number> {
+    const orgId = messages[0]?.organizationId
+    if (!orgId) return 0
+    const ctx = await this.resolveCtx(orgId)
+    return batchStoreMessages(ctx, messages, { batchId, isInitialSync })
+  }
 
-  /**
-   * Deletes messages by their provider external IDs, scoped to an integration.
-   * Cascading FKs handle MessageParticipant cleanup.
-   * Updates thread metadata for affected threads and removes empty threads.
-   * @param integrationId - The integration the messages belong to
-   * @param externalIds - Provider-side message IDs to delete
-   * @returns Number of messages deleted
-   */
   async deleteMessagesByExternalIds(integrationId: string, externalIds: string[]): Promise<number> {
-    if (externalIds.length === 0) return 0
-
-    // Find messages and their threads before deleting
-    const messages = await db
-      .select({ id: schema.Message.id, threadId: schema.Message.threadId })
-      .from(schema.Message)
-      .where(
-        and(
-          eq(schema.Message.integrationId, integrationId),
-          inArray(schema.Message.externalId, externalIds)
-        )
-      )
-
-    if (messages.length === 0) return 0
-
-    const messageIds = messages.map((m) => m.id)
-    const affectedThreadIds = [...new Set(messages.map((m) => m.threadId).filter(Boolean))]
-
-    // Delete messages (cascades to MessageParticipant)
-    await db.delete(schema.Message).where(inArray(schema.Message.id, messageIds))
-
-    logger.info('Deleted messages by external IDs', {
-      integrationId,
-      deletedCount: messageIds.length,
-      affectedThreads: affectedThreadIds.length,
-    })
-
-    // Update or remove affected threads
-    for (const threadId of affectedThreadIds) {
-      if (!threadId) continue
-
-      const [remaining] = await db
-        .select({ count: sql<number>`count(*)::int` })
-        .from(schema.Message)
-        .where(eq(schema.Message.threadId, threadId))
-
-      if (remaining.count === 0) {
-        // Thread is empty — delete it
-        await db.delete(schema.Thread).where(eq(schema.Thread.id, threadId))
-        logger.debug('Deleted empty thread after message removal', { threadId })
-      } else {
-        // Thread still has messages — update metadata
-        await this.updateThreadMetadataEfficient(threadId)
-      }
-    }
-
-    return messageIds.length
+    const ctx = await this.resolveCtx(this.defaultOrganizationId)
+    return deleteMessagesByExternalIds(ctx, { integrationId, externalIds })
   }
 
-  /**
-   * Creates a contact for a participant after sending them a message.
-   * Used when we send a message to someone who previously only sent us messages.
-   * @param participantId The ID of the participant
-   * @param organizationId The ID of the organization
-   */
   async createContactAfterOutboundMessage(
     participantId: string,
     organizationId: string
   ): Promise<void> {
-    try {
-      const [participant] = await db
-        .select()
-        .from(schema.Participant)
-        .where(eq(schema.Participant.id, participantId))
-        .limit(1)
-
-      if (!participant) {
-        logger.warn(`Participant ${participantId} not found for retroactive contact creation`)
-        return
-      }
-
-      // Skip if already linked to a contact
-      if (participant.entityInstanceId) {
-        return
-      }
-
-      // Create contact for this participant
-      const entityInstanceId = await this.findOrCreateContactForParticipant(
-        participant,
-        organizationId,
-        { isInbound: false, role: ParticipantRoleEnum.TO }
-      )
-
-      if (entityInstanceId) {
-        await db
-          .update(schema.Participant)
-          .set({
-            entityInstanceId,
-            hasReceivedMessage: true,
-            lastSentMessageAt: new Date(),
-            updatedAt: new Date(),
-          })
-          .where(eq(schema.Participant.id, participantId))
-
-        logger.info(
-          `Created retroactive contact ${entityInstanceId} for participant ${participantId}`
-        )
-      }
-    } catch (error) {
-      logger.error('Error creating retroactive contact:', {
-        error,
-        participantId,
-        organizationId,
-      })
-    }
+    const ctx = await this.resolveCtx(organizationId)
+    return createContactAfterOutboundMessage(ctx, participantId)
   }
 
-  /**
-   * Ensures contacts exist for all recipients when sending messages in selective mode.
-   * @param recipients Array of email/phone identifiers
-   * @param organizationId The organization ID
-   * @param integrationId The integration ID
-   */
   async ensureContactsForRecipients(
     recipients: string[],
     organizationId: string,
     integrationId: string
   ): Promise<void> {
-    const mode = this.integrationSettings?.recordCreation?.mode || 'selective' // Default to selective mode
-
-    if (mode !== 'selective') {
-      return // Only needed in selective mode
-    }
-
-    for (const identifier of recipients) {
-      if (!identifier) continue
-
-      const identifierType = await this.determineIdentifierType(identifier, integrationId)
-      const normalizedId = this.normalizeIdentifier(identifier, identifierType)
-
-      // Find or create participant
-      const [participant] = await db
-        .select()
-        .from(schema.Participant)
-        .where(
-          and(
-            eq(schema.Participant.organizationId, organizationId),
-            eq(schema.Participant.identifier, normalizedId),
-            eq(schema.Participant.identifierType, identifierType as any)
-          )
-        )
-        .limit(1)
-
-      if (participant && !participant.entityInstanceId) {
-        // Create contact for this participant since we're sending to them
-        await this.createContactAfterOutboundMessage(participant.id, organizationId)
-      }
-    }
+    const ctx = await this.resolveCtx(organizationId)
+    return ensureContactsForRecipients(ctx, { recipients, integrationId })
   }
-} // End of MessageStorageService class
+
+  async getThread(threadId: string, organizationId: string): Promise<Thread | null> {
+    const ctx = await this.resolveCtx(organizationId)
+    return getThread(ctx, { threadId, organizationId })
+  }
+
+  async getThreadMessages(threadId: string): Promise<Message[]> {
+    const ctx = await this.resolveCtx(this.defaultOrganizationId)
+    return getThreadMessages(ctx, threadId)
+  }
+
+  private async resolveCtx(organizationId?: string): Promise<IngestContext> {
+    const orgId = organizationId ?? this.defaultOrganizationId
+    if (!orgId) {
+      throw new Error(
+        'MessageStorageService requires an organizationId — pass one to the constructor or include it on the message.'
+      )
+    }
+    const cached = this.ctxByOrg.get(orgId)
+    if (cached) return cached
+    const promise = createIngestContext(orgId, {
+      integrationSettings: this.integrationSettings,
+      isInitialSync: this.isInitialSync,
+    })
+    this.ctxByOrg.set(orgId, promise)
+    return promise
+  }
+}
